fix examples

+++ examples/FreeRTOS_k1921vk01t_led_blink/src/main.c
@@ -53,7 +53,7 @@ int main()
     periph_init();
     freertos_init_objects();
     printf("K1921VK01T> All periph inited\n");
-    printf("K1921VK01T> CPU frequency is %.3f MHz\n", SystemCoreClock / 1E6);
+    printf("K1921VK01T> CPU frequency is %lu MHz\n", SystemCoreClock / (int)1E6);
 
     vTaskStartScheduler();
     return 0;
